test nogo/coverage: test generated code
This add a test that demonstrate the current behavior of nogo today:
code which was generated under "bazel coverage" to capture coverage
information is going through static analysis.

This behavior is not ideal as end users have very little control over
the way rules_go generated these code and would not be able to fix them
to please the analyzer.

In a future patch, we will introduce a mechanism to alleviate this pain
and flip the expectation in this test.