stdlib: distinct GoStdlib actions
GoStdlib actions provide a set of compiled archives of Go's standard
library for downstream binaries to link against.  In order to support
multiple different compilation and link modes that rules_go users might
use, GoStdlib needs to be prepared in a combination of different
settings.  These combinations include, but not limited to, the
following settings:

- Different toolchains for different OS / CPU architectures
- With / without CGO enabled with different link modes
- With / without race detection enabled

Each unique combination that is dependent upon will generate a new
GoStdlib action in the build.  These actions are displayed to the user's
Bazel build progress outputs like this:

  ```
  GoStdlib stdlib_/pkg; 3s darwin-sandbox
  GoStdlib stdlib_/pkg; 0s darwin-sandbox
  GoStdlib stdlib_/pkg; 1s darwin-sandbox
  ```

This confuses users who are not familiar with rules_go internal into
thinking that the same action are being wrongly executed multiple times
wastefully.  As the result, troubleshooting the Action Graph became
harder and one must rely on the detail outputs in the Build Events
outputs to tell the differences between each actions.

Let's make the mnemonic of GoStdlib actions a bit more descriptive
regarding what's it's doing by including some of the settings mentioned
above into the mnemonic itself.  As the result, we should be able to
see more distinct actions in the Bazel build outputs depending on which
setting was used.

As a result, the actions mnemonic can be distinctly recognized:

  ```
  # Build and output Build Event JSON file
  rules_go> bazel clean
  rules_go> bazel build --build_event_json_file=bes.json --build_event_publish_all_actions //...
  rules_go> grep 'GoStdlib' bes.json |\
              jq 'select(.id | keys[0] == "actionCompleted") | .action.type + " " + .action.label' |\
              sort |\
              uniq -c |\
              sort -n

  # Before this patch
     1 "GoStdlibList //:stdlib"
    14 "GoStdlib //:stdlib"

  # After this patch
    1 "GoStdlibDarwinAmd64 //:stdlib_pure"
    1 "GoStdlibDarwinArm64CGoLinkCArchive //:stdlib_cgo"
    1 "GoStdlibDarwinArm64CGoLinkCShared //:stdlib_cgo"
    1 "GoStdlibDarwinArm64CGoLinkPie //:stdlib_cgo"
    1 "GoStdlibDarwinArm64CGoLinkPlugin //:stdlib_cgo"
    1 "GoStdlibLinuxAmd64 //:stdlib_pure"
    1 "GoStdlibList //:stdlib_cgo"
    1 "GoStdlibPlan9Amd64 //:stdlib_pure"
    1 "GoStdlibPlan9Arm //:stdlib_pure"
    1 "GoStdlibWindowsAmd64 //:stdlib_pure"
    2 "GoStdlibDarwinArm64RaceCGo //:stdlib_cgo"
    3 "GoStdlibDarwinArm64 //:stdlib_cgo"
  ```

Note how this still not yet fully de-duplicate the mnemonic.  This is
likely due to the fact that different version of `builder` binaries was
used. But it's unclear to me why that is and whether these actions are
truly duplicated?

  ```
  > grep 'GoStdlib' bes.json |\
      jq 'select(.id | keys[0] == "actionCompleted") |\
            select( .action.type == "GoStdlibDarwinArm64") |\
            select(.action.label == "//:stdlib_cgo") |\
            .action.commandLine[0]'
  "bazel-out/darwin_arm64-opt-exec-2B5CBBC6-ST-36745079df94/bin/external/go_sdk/builder"
  "bazel-out/darwin_arm64-opt-exec-2B5CBBC6-ST-6ca1a6e450d1/bin/external/go_sdk/builder"
  "bazel-out/darwin_arm64-opt-exec-2B5CBBC6-ST-149cbabf6624/bin/external/go_sdk/builder"
  ```