upgrade x/tools to v0.12.0
Upgrade was done with

```bash
> bazel run go/tools/releaser -- upgrade-dep org_golang_x_tools
```

after the patch in #3677.

---

A few additional updates needed for this to work:

- Update to `tests/integration/popular_repos` to remove unused packages
- Update to `nogo` to use the new interface of x/tools

Because this upgrade is NOT backward compatible (by design of x/tools
repo), downstream users of rules_go are advised to upgrade `x/tools`
version in their repo to v0.12.0 or later to avoid breakages.